+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 0.0.4
 - Expanded pipeline discovery to support `*pipeline*.py` patterns and improved handling of nested subdirectories.
+- Add new extension setting `Kedro: Environment` to change the configuration source.
 
 # 0.0.3
 - Add catalog config preview for mouse hover.