+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Add catalog config preview for mouse hover.
 
 # 0.0.2
-- Fix a startup error bug.
+- Fix a **startup** error bug.
 
 # 0.0.1
 - Support basic go to definition