+++ CHANGELOG.md
@@ -2,13 +2,14 @@
 
 # 0.0.4
 - Expanded pipeline discovery to support `*pipeline*.py` patterns and improved handling of nested subdirectories.
-- Add new extension setting `Kedro: Environment` to change the configuration source.
+- Add new extension setting `Kedro: Environment` to change the configuration environment.
+- Add new command `kedro: Select Environment` to change the configuration environment.
 
 # 0.0.3
 - Add catalog config preview for mouse hover.
 
 # 0.0.2
-- Fix a startup error bug.
+- Fix a **startup** error bug.
 
 # 0.0.1
 - Support basic go to definition