+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Change Log
 
-# 0.0.4
+# 0.1.0
 - Expanded pipeline discovery to support `*pipeline*.py` patterns and improved handling of nested subdirectories.
 - Add new extension setting `Kedro: Environment` to change the configuration environment.
 - Add new command `kedro: Select Environment` to change the configuration environment.
 - `Go to Definition` now search `default_run_env` first.
+- Add Kedro Icon and Status bar to select environment.
+
 
 # 0.0.3
 - Add catalog config preview for mouse hover.